requirements.txt
langchain
langchain-community
faiss-cpu
langchain-ollama
python-dotenv
docling
langchain-docling

openai>=1.55.3
PyPDF2 
sentence-transformers 
numpy 
tqdm
huggingface-hub
pandas
datasets
pycryptodome

ragas
langchain-openai
langchain
torch
transformers